refactor(api): move text processing logic out of router
